backend/requirements.txt
fastapi
uvicorn
google-generativeai
pydantic
requests
pyyaml
python-multipart
python-dotenv
